pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["brunoqs <brunoqueiroz020@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
sqlalchemy = "^2.0.27"
uvicorn = "^0.27.1"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"


[tool.poetry.group.test.dependencies]
pytest = "^8.0.2"
httpx = "^0.27.0"


[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
mypy = "^1.8.0"
flake8 = "^7.0.0"
isort = "^5.13.2"
pylint = "^3.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
